requirements.txt
streamlit
langchain
chromadb
google-generativeai
tiktoken
numpy
pandas